refactor: server action이 모인 파일명을 actions.ts로 변경
